split refresh into own module
